% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_me_tidy.R
\name{hello_world}
\alias{hello_world}
\title{Prints "Hello World" to see if github actions work}
\usage{
hello_world()
}
\value{
Nothing
}
\description{
Prints "Hello World" to see if github actions work
}
\examples{
hello_world()
}
