% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetreg.R
\name{hetreg}
\alias{hetreg}
\title{Heteroskedastic linear regression}
\usage{
hetreg(data, y, x, het = x, ..., gls_opt = NULL)
}
\arguments{
\item{data}{data frame containing the variables in \code{y}, \code{x} and \code{het}.}

\item{y}{string (character vector) with the name of the independent variable.}

\item{x}{vector of strings (character vectors) with the name of the independent variable(s).}

\item{het}{vector of strings (character vectors) with the name of the variable(s) that enter(s) the variance function as independent variable(s).}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}.}

\item{gls_opt}{optional arguments passed to \code{\link[nlme:gls]{gls()}}.}
}
\value{
an object of class \code{gls} representing the linear model fit. See \code{\link[nlme:gls]{gls()}} for details.
}
\description{
\code{hetreg()} fits a heteroskedastic linear regression
where the variance is an exponential function of specified variables
}
\details{
\code{hetreg()} is a wrapper for \code{\link[nlme:gls]{nlme::gls()}}. Compared to \code{\link[nlme:gls]{gls()}},
the function \code{hetreg()}
\itemize{
\item is pipeable (plays nice with the tidyverse)
\item automatically constructs the \code{\link[nlme:gls]{gls()}} argument \code{model} from
\code{y} and \code{x}, and the \code{\link[nlme:gls]{gls()}} argument \code{weights} from \code{het}
\item only allows for an exponential variance function structure defined by
a combination \code{\link[nlme:varComb]{varComb()}} of \link[nlme:varExp]{varExp()} terms
\item has tidy support
}

The model estimated is
\deqn{y_i = x_i \beta + \epsilon_i \\
      var(\epsilon_i) = \exp(2 \theta v_i )}
where where \eqn{y_i}, \eqn{i = 1,\dotsc,n}, is the dependent variable;
\eqn{x_i = (x_{1i},x_{2i},\dotsc,x_{ki})} are the k independent
variables that model the mean function; \eqn{v_i = (v_{1i},v_{2i},\dotsc,v_{mi})}
are the m variables that model the variance function (corresponding to the
argument \code{het}); \eqn{\beta} is a column vector of unknown parameters in the
mean function, and \eqn{\theta} is a column vector of unknown
parameters in the variance function.
\code{hetreg()} is similar to the command \href{https://www.stata.com/manuals/rhetregress.pdf}{hetregress}
in Stata, where

\if{html}{\out{<div class="sourceCode Stata">}}\preformatted{hetregress x y, het(z)
}\if{html}{\out{</div>}}

corresponds to the model
\deqn{y_i = x_i \beta + \epsilon_i \\
      var(\epsilon_i) = \exp(z_i \alpha )}
The \eqn{z_i} map to \eqn{v_i} and \eqn{\alpha} maps to \eqn{2 \theta}.
}
