% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfci_hetreg_tidiers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{augment}
\title{Objects exported from other packages}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about model components.
}
\seealso{
\href{https://github.com/bbolker/broom.mixed}{\code{broom.mixed::tidy.gls()}}

\href{https://github.com/bbolker/broom.mixed}{\code{broom.mixed::glance.gls()}}

\href{https://github.com/bbolker/broom.mixed}{\code{broom.mixed::augment.gls()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
}}

