% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfci_hetreg_tidiers.R
\name{tidy.het}
\alias{tidy.het}
\alias{het_tidiers}
\alias{glance.het}
\alias{augment.het}
\title{Tidying methods for heteroskedastic linear regressions}
\usage{
\method{tidy}{het}(x, conf_int = FALSE, conf_level = 0.95, ...)

\method{glance}{het}(x, ...)

\method{augment}{het}(x, ...)
}
\arguments{
\item{x}{An object of class \code{het} created by \code{\link[=hetreg]{hetreg()}}.}

\item{conf_int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf_level}{The confidence level to use for the confidence interval
if \code{conf_int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}.}
}
\value{
Various statistics and post-estimation values for \code{\link[=hetreg]{hetreg()}}
}
\description{
Tidying methods \code{tidy}, \code{glance} and \code{augment} for objects of class \code{het}
that are compatible with \link[broom:broom]{broom}.
}
\section{Functions}{
\itemize{
\item \code{tidy(het)}: returns the same information as \code{nlme_tidiers()} from
\href{https://github.com/bbolker/broom.mixed}{\code{broom.mixed}}
for the mean equation, and adds coefficients, standard errors, t-statistics
and p-values for the variance equation

\item \code{glance(het)}: same as
\href{https://github.com/bbolker/broom.mixed}{\code{broom.mixed::glance.gls()}}

\item \code{augment(het)}: returns the standard deviation of the error term

}}
\seealso{
\code{\link[broom:reexports]{broom::tidy()}}, \code{nlme_tidiers()} from
\href{https://github.com/bbolker/broom.mixed}{\code{broom.mixed}}
}
