% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfci-package.R
\docType{package}
\name{vfci-package}
\alias{vfci}
\alias{vfci-package}
\title{vfci: Code for "The Market Price of Risk and Macro-Financial Dynamics"}
\description{
R package to reproduce tables and figures in Adrian, Duarte and Iyer (2023), “The Market Price of Risk and Macro-Financial Dynamics”, [CEPR Press Discussion Paper No. 17777](https://cepr.org/publications/dp17777).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VFCI/vfci}
  \item Report bugs at \url{https://github.com/VFCI/vfci/issues}
}

}
\author{
\strong{Maintainer}: Tara Iyer \email{taraiyer8@gmail.com}

Authors:
\itemize{
  \item Tobias Adrian \email{tobiasadrian@gmail.com} (\href{https://orcid.org/0000-0001-9379-9592}{ORCID})
  \item Fernando Duarte \email{fernando.duarte@gmail.com} (\href{https://orcid.org/0000-0002-8495-6936}{ORCID})
}

}
\keyword{internal}
